% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_contributions.R
\name{contributions_calculation}
\alias{contributions_calculation}
\title{Calculate study contributions}
\usage{
contributions_calculation(data)
}
\arguments{
\item{data}{Data frame in the format of \code{data_genes}}
}
\value{
A tibble
}
\description{
Calculates the raw and relative information an contribution of studies to MAIC.
}
\details{
Input columns for \code{data_genes} should be (this is the standard output of the MAIC algorithm):
\itemize{
\item \code{gene} - HGNC gene symbol - chr
\item 1...
\item \code{uID} - Study unique identifier. Column contains study specific gene score - dbl
\item n...
\item \code{maic_score} - MAIC score for gene - dbl
\item \code{contributors} - Studies contributing to MAIC score by method - chr
}

Definitions:
\itemize{
\item \code{raw_information} = sum of gene scores for each study
\item \code{relative_information} = \code{raw_information}/sum of all gene scores in MAIC
\item \code{raw_contribution} = sum of gene scores for each study which contribute to the MAIC score
\item \code{relative_contribution} = \code{raw_contribution}/sum of all gene scores in MAIC contributing to the
MAIC score
}
}
\examples{
\dontrun{
if(interactive()){
 contributions_calculation(data_genes)
 }
}
}
