% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_contributions_chord_diagrams.R
\name{contributions_chord_bymethod}
\alias{contributions_chord_bymethod}
\title{Create a chord diagram of contributions by method type}
\usage{
contributions_chord_bymethod(data_contributions, data_study)
}
\arguments{
\item{data_contributions}{-- Input in the format of \code{data_contributions}}

\item{data_study}{-- Input in the format of \code{data_study}}
}
\value{
An html chord plot
}
\description{
Creates an interactive html plot of contributions by method type
}
\examples{
\dontrun{
if(interactive()){
 contributions_chord_bymethod(data_contributions, data_study)
 }
}
}
