% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_methods_and_gene_support.R
\name{count_genes}
\alias{count_genes}
\title{Count genes}
\usage{
count_genes(data)
}
\arguments{
\item{data}{Data frame in the format of \code{data_genes}}
}
\value{
An integer
}
\description{
Returns the total number of unique genes or SNPs included in MAIC.
}
\details{
Input columns for \code{data_genes} should be (this is the standard output of the MAIC algorithm):
\itemize{
\item \code{gene} - HGNC gene symbol - chr
\item 1...
\item \code{uID} - Study unique identifier. Column contains study specific gene score - dbl
\item n...
\item \code{maic_score} - MAIC score for gene - dbl
\item \code{contributors} - Studies contributing to MAIC score by method - chr
}
}
\examples{
\dontrun{
if(interactive()){
 count_genes(ARDSMAICR::data_genes)
 }
}
}
