% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_enrichment.R
\name{count_gp_pathways}
\alias{count_gp_pathways}
\title{Count the number of enriched pathways}
\usage{
count_gp_pathways(gp_result, by_pathway = FALSE)
}
\arguments{
\item{gp_result}{-- The return object of gp_enrichment()}

\item{by_pathway}{-- Count by pathway database -- Default: FALSE}
}
\value{
Either an integer or a tibble
}
\description{
Returns the total number of enriched pathways or the number of enriched pathways per
database
}
\examples{
\dontrun{
if(interactive()){
 data_genes |>
   gp_enrichment(0.01, c("WP", "HPA")) |>
   count_gp_pathways(by_pathway = TRUE)
 }
}
}
