% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_methods_and_gene_support.R
\name{count_lists}
\alias{count_lists}
\title{Count gene lists}
\usage{
count_lists(data)
}
\arguments{
\item{data}{-- Input in the format of \code{data_study}}
}
\value{
An integer
}
\description{
Return the total number of included gene lists
}
\examples{
\dontrun{
if(interactive()){
 count_lists(data_study)
 }
}
}
