% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ards_literature_overlap.R
\name{count_overlap}
\alias{count_overlap}
\title{Count MAIC and alternative gene list overlap}
\usage{
count_overlap(data_genes, data_alternative)
}
\arguments{
\item{data_genes}{-- Input in the format of \code{data_genes}}

\item{data_alternative}{-- Dataframe where the first column is a gene list}
}
\value{
A list
}
\description{
Return the total number of genes identified in the alternative list, the number of
these genes found in ARDS MAIC, and the percentage of alternative list genes found in ARDS MAIC.
}
\examples{
\dontrun{
if(interactive()){
 count_overlap(data_genes, data_biolitmine)
 }
}
}
