% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_biolitmine}
\alias{data_biolitmine}
\title{ARDS MAIC Update 1.0.0 BioLitMine results}
\format{
\subsection{\code{data_biolitmine}}{

A data frame with 271 rows and 3 columns:
\describe{
\item{Gene}{HGNC symbol}
\item{Last_10_PMIDS}{The PubMed IDs of the last 10 associated publications}
\item{Publication_count}{The total number of associated publications}
...
}
}
}
\source{
\url{https://www.flyrnai.org/tools/biolitmine/web/}
}
\usage{
data_biolitmine
}
\description{
Genes associated with the MeSH heading 'Respiratory Distress syndrome, Acute' obtained on the
4th January, 2023
}
\keyword{datasets}
