% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_contributions}
\alias{data_contributions}
\title{Overall gene list contributions - ARDS MAIC Update 1.0.0}
\format{
\subsection{\code{data_contributions}}{

A data frame with 43 rows and 5 columns:
\describe{
\item{study}{Gene list}
\item{raw_information}{Raw information}
\item{information}{Information}
\item{raw_contribution}{Raw contribution}
\item{contribution}{Contribution}
...
}
}
}
\source{
Own
}
\usage{
data_contributions
}
\description{
Gene list information content and information contribution
}
\keyword{datasets}
