% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_covidmaic}
\alias{data_covidmaic}
\title{SARS-CoV-2 MAIC genes - ARDS MAIC Update 1.0.0}
\format{
\subsection{\code{data_covidmaic}}{

A data frame with 5418 rows and 1 column:
\describe{
\item{gene}{HGNC symbol}
...
}
}
}
\source{
\url{https://doi.org/10.1038/s41598-020-79033-3}
}
\usage{
data_covidmaic
}
\description{
Genes included in a MAIC of SARS-CoV-2 genome-wide studies
}
\keyword{datasets}
