% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_genes}
\alias{data_genes}
\title{ARDS MAIC Update 1.0.0 overall maic results}
\format{
\subsection{\code{data_genes}}{

A data frame with 7088 rows and 46 columns:
\describe{
\item{gene}{HGNC symbol}
\item{contributors}{Studies contributing to MAIC score}
...
}
}
}
\source{
Own
}
\usage{
data_genes
}
\description{
MAIC output for all gene lists
}
\keyword{datasets}
