% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_study}
\alias{data_study}
\title{Study metadata - ARDS MAIC Update 1.0.0}
\format{
\subsection{\code{data_study}}{

A data frame with 44 rows and 16 columns:
\describe{
\item{id}{integer}
\item{First_author}{First author's surname}
\item{Article_title}{Title}
\item{Year}{Publication year}
\item{Journal}{Journal}
\item{DOI}{Digital Object Identifier}
\item{PMID}{Pubmed ID}
\item{uID}{Unique identifier}
\item{Method}{Method used to create list}
\item{Technology}{Specific technology used}
\item{Tissue}{Tissue samples derived from}
\item{Cell}{Cell type sampled}
\item{Focus}{Study objective}
\item{ARDS_pts}{Total number of ARDS patients included in study}
\item{ARDS_definition}{ARDS definition used in study}
\item{List_available}{Is a gene list retrievable}
...
}
}
}
\source{
Own
}
\usage{
data_study
}
\description{
A description of each gene list included in the systematic review and ARDS MAIC
}
\keyword{datasets}
