% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_and_methods_tables.R
\name{design_table}
\alias{design_table}
\title{Create a table of study designs}
\usage{
design_table(data)
}
\arguments{
\item{data}{-- Input in the format of \code{data_study}}
}
\value{
An html table
}
\description{
Create an interactive html table of study designs
}
\examples{
\dontrun{
if(interactive()){
 design_table(data_study)
 }
}
}
