% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_table.R
\name{gene_table}
\alias{gene_table}
\title{Create gene table}
\usage{
gene_table(data)
}
\arguments{
\item{data}{-- Input in the format of \code{data_genes}}
}
\value{
An html table
}
\description{
Create a searchable html table of genes ranked by maic score
}
\examples{
\dontrun{
if(interactive()){
 gene_table(data_genes)
 }
}
}
