% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_methods_and_gene_support.R
\name{genes_per_method}
\alias{genes_per_method}
\title{Count genes per method}
\usage{
genes_per_method(data_study, data_genes)
}
\arguments{
\item{data_study}{-- Input in the format of \code{data_study}}

\item{data_genes}{-- Input in the format of \code{data_genes}}
}
\value{
A tibble
}
\description{
Return the number of unique genes or SNPs identified by each method type
}
\examples{
\dontrun{
if(interactive()){
 genes_per_method(data_study, data_genes)
 }
}
}
