% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_enrichment.R
\name{gp_count_pathways}
\alias{gp_count_pathways}
\title{Count enriched pathways}
\usage{
gp_count_pathways(gp_result, by_pathway = FALSE)
}
\arguments{
\item{gp_result}{The return object of \code{gp_enrichment()}}

\item{by_pathway}{Count by pathway database -- Default = FALSE}
}
\value{
Either an integer or a tibble
}
\description{
Returns the total number of enriched pathways or the number of enriched pathways per
database.
}
\examples{
\dontrun{
if(interactive()){
 gp_enrichment(data_genes, p_threshold = 0.001, source = c("REAC", "WP")) |>
 gp_count_pathways()
 }
}
}
