% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_enrichment.R
\name{gp_enrichment}
\alias{gp_enrichment}
\title{Functional enrichment using gprofiler}
\usage{
gp_enrichment(
  data,
  p_threshold = 0.01,
  source = c("GO", "KEGG", "REAC", "WP", "HPA")
)
}
\arguments{
\item{data}{-- Input in the format of \code{data_study}}

\item{p_threshold}{-- Threshold P value -- Default: 0.01}

\item{source}{-- Database -- any combination of GO, GO:MF, GO:CC, GO:BP, KEGG, REAC, WP, or HPA
-- Default: c("GO", "KEGG", "REAC", "WP", "HPA")}
}
\value{
Enrichment result and associated metadata
}
\description{
Performs functional enrichment of standard maic output against a range of databases
}
\details{
GO = Gene Ontology (MF = molecular function; CC = cellular component,
BP = biological process), KEGG = Kyoto Encyclopedia of Genes and Genomes, REAC = Reactome,
WP = WikiPathways, HPA = Human Protein Atlas
}
\examples{
\dontrun{
if(interactive()){
 gp_enrichment(data_genes, 0.001, c("GO", "WP"))
 }
}
}
