% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_enrichment.R
\name{gp_plot}
\alias{gp_plot}
\title{Functional enrichment plot}
\usage{
gp_plot(gp_result, is_interactive = TRUE)
}
\arguments{
\item{gp_result}{-- The return object of gp_enrichment()}

\item{is_interactive}{-- Interactive plot -- Default: TRUE}
}
\value{
A Manhattan plot
}
\description{
Create a Manhattan plot of functional enrichment results
}
\examples{
\dontrun{
if(interactive()){
 data_genes |>
   gp_enrichment(0.001, "WP") |>
   gp_plot(is_interactive = FALSE)
 }
}
}
