% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_enrichment.R
\name{gp_table}
\alias{gp_table}
\title{Functional enrichment results table}
\usage{
gp_table(gp_tidy_result)
}
\arguments{
\item{gp_tidy_result}{The return object of \code{gp_tidy_results()}}
}
\value{
An html table
}
\description{
Creates an interactive html table of the tidy output of gprofiler.
}
\examples{
\dontrun{
if(interactive()){
 gp_enrichment(ARDSMAICR::data_genes, p_threshold = 0.001, source = c("REAC", "WP")) |>
 gp_tidy_results() |>
 gp_table()
 }
}
}
