% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_enrichment.R
\name{gp_tidy_results}
\alias{gp_tidy_results}
\title{Tidy functional enrichment results}
\usage{
gp_tidy_results(gp_result)
}
\arguments{
\item{gp_result}{The return object of \code{gp_enrichment()}}
}
\value{
A tibble
}
\description{
Returns a tidy tibble of the gprofiler output.
}
\examples{
\dontrun{
if(interactive()){
 gp_enrichment(data_genes, p_threshold = 0.001, source = c("REAC", "WP")) |>
 gp_tidy_results()
 }
}
}
