% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_enrichment.R
\name{gp_table}
\alias{gp_table}
\title{Create a table of functional enrichment results}
\usage{
gp_table(tidy_gp_res)
}
\arguments{
\item{tidy_gp_res}{-- The return object of tidy_gp_results()}
}
\value{
An html table
}
\description{
Creates an interactive html table of the tidy output of gprofiler
}
\examples{
\dontrun{
if(interactive()){
 data_genes |>
   gp_enrichment(0.001, "WP") |>
   tidy_gp_results() |>
   gp_table()
 }
}
}
