% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC_adjacency_matrix.R
\name{inf_adj_matrix}
\alias{inf_adj_matrix}
\title{Information adjacency matrix}
\usage{
inf_adj_matrix(data, contribution = FALSE, unique = TRUE, as_list = FALSE)
}
\arguments{
\item{data}{Data frame in the format of \code{ARDSMAICR::data_genes}}

\item{contribution}{Return the information contribution -- Default = FALSE}

\item{unique}{Include information for genes uniquely identified by a study -- Default = TRUE}

\item{as_list}{Option to return an adjacency list -- Default = FALSE}
}
\value{
A matrix with the dimensions n studies x n studies. Where the cells represent the summed
scores of genes shared between studies. Direction is maintained, columns to rows represents
the information in the column shared with the row. The return may be specified as the
information content - all gene scores regardless of their contribution to the MAIC score - or
the information contribution - only gene scores where the study contributes to the MAIC score
for that gene. The option to include or remove information content/contribution generated by
a study uniquely is available.
}
\description{
Creates an adjacency matrix of shared information content/contribution between
studies.
}
\details{
Input columns for \code{data_genes} should be (this is the standard output of the MAIC algorithm):
\itemize{
\item \code{gene} - HGNC gene symbol - chr
\item 1...
\item \code{uID} - Study unique identifier. Column contains study specific gene score - dbl
\item n...
\item \code{maic_score} - MAIC score for gene - dbl
\item \code{contributors} - Studies contributing to MAIC score by method - chr
}
}
\examples{
\dontrun{
if(interactive()){
 inf_agj_matrix(ARDSMAICR::data_genes, contribution = TRUE, unique = FALSE)
 }
}
}
