% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_prioritisation.R
\name{inflection_point}
\alias{inflection_point}
\title{Calculate the maic score inflection point}
\usage{
inflection_point(data)
}
\arguments{
\item{data}{-- Input in the format of \code{data_genes}}
}
\value{
A list of values
}
\description{
Calculates the maic score and gene number at which the inflection point in the curve
of genes ranked by maic score occurs
}
\examples{
\dontrun{
if(interactive()){
 inflection_point(data_genes)
 }
}
}
\references{
\url{https://CRAN.R-project.org/package=inflection}
}
