% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_protein_interaction.R
\name{initialise_string_db}
\alias{initialise_string_db}
\title{Initialise STRING db}
\usage{
initialise_string_db(
  vers = "11.5",
  score_threshold = 200,
  input_directory = ""
)
}
\arguments{
\item{vers}{-- STRING dbVersion -- Default: '11.5'}

\item{score_threshold}{-- Score threshold --  Default: 200}

\item{input_directory}{-- Input directory (can save for later use) -- Default: ''}
}
\value{
An S4 object
}
\description{
Initialises STRING db using the Bioconducter package \code{STRINGdb} and saves to a temp
directory
}
\details{
Creates a full network
}
\examples{
\dontrun{
if(interactive()){
 initialise_string_db()
 }
}
}
