% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_methods_and_gene_support.R
\name{lists_per_gene}
\alias{lists_per_gene}
\title{Count the number of lists each gene or SNP is found in}
\usage{
lists_per_gene(data)
}
\arguments{
\item{data}{-- Input in the format of \code{data_genes}}
}
\value{
A tibble
}
\description{
For each unique gene or SNP return the number of lists that gene or SNP is
found in
}
\examples{
\dontrun{
if(interactive()){
 IL6_list_count <- data_genes |>
   lists_per_gene() |>
   dplyr::filter(.data$gene == "IL6")
 print(IL6_list_count)
 }
}
}
