% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_methods_and_gene_support.R
\name{lists_per_method}
\alias{lists_per_method}
\title{Count gene lists per method type}
\usage{
lists_per_method(data)
}
\arguments{
\item{data}{-- Input in the format of \code{data_study}}
}
\value{
A tibble
}
\description{
Return the total number of gene lists belonging to each method identified
}
\examples{
\dontrun{
if(interactive()){
 lists_per_method(data_study)
 }
}
}
