% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_methods_and_gene_support.R
\name{lists_per_method_per_gene}
\alias{lists_per_method_per_gene}
\title{Count the number of lists per method for each gene or SNP}
\usage{
lists_per_method_per_gene(data_study, data_genes)
}
\arguments{
\item{data_study}{-- Input in the format of \code{data_study}}

\item{data_genes}{-- Input in the format of \code{data_genes}}
}
\value{
A tibble
}
\description{
Return the number of lists each unique gene or SNP is found in stratified
by method type
}
\examples{
\dontrun{
if(interactive()){
 IL6_list_per_method_count <- data_genes |>
   lists_per_method_per_gene() |>
   dplyr::filter(.data$gene == "IL6")
 print(IL6_list_per_method_count)
 }
}
}
