% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_protein_interaction.R
\name{map_string_db}
\alias{map_string_db}
\title{Map genes to STRING IDs}
\usage{
map_string_db(string_db, data_genes, n_genes = NULL)
}
\arguments{
\item{string_db}{-- Object output from \code{initialise_string_db}}

\item{data_genes}{-- Input in the format of \code{data_genes}}

\item{n_genes}{P-- Number of genes to map from highest ranked maic gene to n}
}
\value{
A tibble
}
\description{
Maps input genes to STRING IDs
}
\examples{
\dontrun{
if(interactive()){
 initialise_string_db()
 mapped_genes <- map_string_db(data_genes, n_genes = 100)
 }
}
}
\seealso{
see the STRINGdb package for further functions to work with the output
}
