% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_methods_and_gene_support.R
\name{methods_per_gene}
\alias{methods_per_gene}
\title{Count the number of method types for each gene}
\usage{
methods_per_gene(data)
}
\arguments{
\item{data}{-- Input in the format of \code{data_genes}}
}
\value{
A tibble
}
\description{
Return the number of method types supporting each unique gene or SNP
}
\examples{
\dontrun{
if(interactive()){
 IL6_method_support <- data_genes |>
   methods_per_gene() |>
   dplyr::filter(.data$gene == "IL6")
 print(IL6_method_support)
 }
}
}
