% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_and_methods_tables.R
\name{methods_table}
\alias{methods_table}
\title{Create a table of study methods}
\usage{
methods_table(data)
}
\arguments{
\item{data}{-- Input in the format of \code{data_study}}
}
\value{
An html table
}
\description{
Create an interactive html table of study methods
}
\examples{
\dontrun{
if(interactive()){
 methods_table(data_study)
 }
}
}
