% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC_adjacency_matrix.R
\name{normalise_gene_scores}
\alias{normalise_gene_scores}
\title{Normalise gene scores}
\usage{
normalise_gene_scores(data)
}
\arguments{
\item{data}{Data frame in the format of \code{data_genes}}
}
\value{
A data frame
}
\description{
Normalises genes scores to account for the length of the input list.
}
\details{
Normalised by dividing the gene scores for each list by the length of the list.
Input columns for \code{data_genes} should be (this is the standard output of the MAIC algorithm):
\itemize{
\item \code{gene} - HGNC gene symbol - chr
\item 1...
\item \code{uID} - Study unique identifier. Column contains study specific gene score - dbl
\item n...
\item \code{maic_score} - MAIC score for gene - dbl
\item \code{contributors} - Studies contributing to MAIC score by method - chr
}
}
\examples{
\dontrun{
if(interactive()){
 normalised_info_content_matrix <- data_genes |>
 normalise_gene_scores() |>
 inf_adj_matrix(unique = FALSE)
 }
}
}
