% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ards_literature_overlap.R
\name{overlap_table}
\alias{overlap_table}
\title{Create a table of ARDS genes found in an alternative list of genes}
\usage{
overlap_table(data_genes, data_alternative, biolitmine = TRUE)
}
\arguments{
\item{data_genes}{-- Input in the format of \code{data_genes}}

\item{data_alternative}{-- Dataframe where the only column is gene symbols and titled gene,
unless is \code{data_biolitmine}}

\item{biolitmine}{-- TRUE or FALSE -- Default = TRUE -- if comparing BioLitMine result using
\code{data_biolitmine}}
}
\value{
An html table
}
\description{
Creates an interactive html table of genes found in common between ARDS MAIC and a
and a user supplied list of genes
}
\details{
Primary use is to compare ARDS MAIC with the results of a BioLitMine search for the MeSH
term "Respiratory Distress Syndrome, Adult". A secondary use case is to compare ARDS MAIC
with MAIC analyses performed in different conditions e.g., our pervious MAIC of SARS-CoV-2
whole-genome studies.
}
\examples{
\dontrun{
if(interactive()){
 overlap_table(data_genes, data_covidmaic, biolitmine = FALSE)
 }
}
}
