% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_overlap.R
\name{overlap_test}
\alias{overlap_test}
\title{Compute measures of gene overlap}
\usage{
overlap_test(data_genes, overlap_genes, n_data_genes = "All", universe = 19220)
}
\arguments{
\item{data_genes}{Data frame in the format of \code{data_study}}

\item{overlap_genes}{Data frame where the first column lists gene symbols}

\item{n_data_genes}{Number of MAIC genes to include -- Default = "All" -- chr}

\item{universe}{Size of gene universe. Either a user defined integer or "HGNC" = All protein
coding genes by HGNC, "FANTOM_L" = genes expressed in lung in FANTOM5, or "FANTOM_S" =
genes expressed in spleen in FANTOM5 -- Default: "HGNC" -- chr}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Compute measures of gene overlap between MAIC genes (or a subset) and a user defined
list of genes.
}
\details{
Input columns for \code{data_genes} should be (this is the standard output of the MAIC algorithm):
\itemize{
\item \code{gene} - HGNC gene symbol - chr
\item 1...
\item \code{uID} - Study unique identifier. Column contains study specific gene score - dbl
\item n...
\item \code{maic_score} - MAIC score for gene - dbl
\item \code{contributors} - Studies contributing to MAIC score by method - chr
\item "HGNC" = All protein coding genes in HGNC = 19220 genes
\item "FANTOM_L" = Genes expressed in lung in FANTOM5 = 12606 genes
\item "FANTOM_S" = Genes expressed in spleen in FANTOM5 = 11512 genes
}
}
\examples{
\dontrun{
if(interactive()){
 overlap_test(data_genes, data_biolitlmine, n_data_genes = 300, universe = "FANTOM_L")
 }
}
}
