% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ards_literature_overlap.R
\name{overlap_unidentified_table}
\alias{overlap_unidentified_table}
\title{MAIC unidentified table}
\usage{
overlap_unidentified_table(data_genes, data_alternative, biolitmine = TRUE)
}
\arguments{
\item{data_genes}{Data frame in the format of \code{ARDSMAICR::data_genes}}

\item{data_alternative}{Data frame where the only column is gene symbols and titled gene,
unless is \code{ARDSMAICR::data_biolitmine}}

\item{biolitmine}{Boolean -- Default = TRUE -- if comparing BioLitMine result using
\code{ARDSMAICR::data_biolitmine}}
}
\value{
An html table (biolitmine) or a tibble (other source)
}
\description{
Create an interactive html table of genes found in an alternative list but not in
MAIC.
}
\details{
Primary use is to compare ARDS MAIC with the results of a BioLitMine search for the MeSH
term "Respiratory Distress Syndrome, Adult". A secondary use case is to compare ARDS MAIC
with MAIC analyses performed in different conditions e.g., our pervious MAIC of SARS-CoV-2
whole-genome studies.

Input columns for \code{data_genes} should be (this is the standard output of the MAIC algorithm):
\itemize{
\item \code{gene} - HGNC gene symbol - chr
\item 1...
\item \code{uID} - Study unique identifier. Column contains study specific gene score - dbl
\item n...
\item \code{maic_score} - MAIC score for gene - dbl
\item \code{contributors} - Studies contributing to MAIC score by method - chr
}
}
\examples{
\dontrun{
if(interactive()){
 overlap_unidentified_table(ARDSMAICR::data_genes, ARDSMAICR::data_biolitmine, biolitmine = TRUE)
 }
}
}
