% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ards_literature_overlap.R
\name{overlap_venn}
\alias{overlap_venn}
\title{MAIC overlap Venn diagram}
\usage{
overlap_venn(data_genes, data_alternative, biolitmine = TRUE)
}
\arguments{
\item{data_genes}{Data frame in the format of \code{ARDSMAICR::data_genes}}

\item{data_alternative}{Data frame where the only column is gene symbols and titled gene,
unless is \code{ARDSMAICR::data_biolitmine}}

\item{biolitmine}{Boolean -- Default = TRUE -- if comparing BioLitMine result using
\code{ARDSMAICR::data_biolitmine}}
}
\value{
Venn diagram
}
\description{
Creates a Venn diagram of the overlap between MAIC genes and genes found in an
alternative list.
}
\details{
Input columns for \code{data_genes} should be (this is the standard output of the MAIC algorithm):
\itemize{
\item \code{gene} - HGNC gene symbol - chr
\item 1...
\item \code{uID} - Study unique identifier. Column contains study specific gene score - dbl
\item n...
\item \code{maic_score} - MAIC score for gene - dbl
\item \code{contributors} - Studies contributing to MAIC score by method - chr
}
}
\examples{
\dontrun{
if(interactive()){
 overlap_venn(ARDSMAICR::data_genes, ARDSMAICR::data_biolitmine, biolitmine = TRUE)
 }
}
}
