% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_overlap.R
\name{test_gene_overlap}
\alias{test_gene_overlap}
\title{Compute measures of gene overlap}
\usage{
test_gene_overlap(
  data_genes,
  overlap_genes,
  n_data_genes = "All",
  universe = 19220
)
}
\arguments{
\item{data_genes}{-- Input in the format of \code{data_study}}

\item{overlap_genes}{-- Input where the first column lists gene symbols}

\item{n_data_genes}{-- Number of maic genes to include -- Default: 'All'}

\item{universe}{-- Size of gene universe. Either a user defined integer or "HGNC" = All protein
coding genes by HGNC, "FANTOM_L" = genes expressed in lung in FANTOM5, or "FANTOM_S" =
genes expressed in spleen in FANTOM5 -- Default: "HGNC"}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Compute measures of gene overlap between maic genes or a subset and a user defined
list of genes
}
\details{
"HGNC" = 19220 genes, "FANTOM_L" = 12606 genes, "FANTOM_S" = 11512 genes
}
\examples{
\dontrun{
if(interactive()){
 test_gene_overlap(data_genes, data_biolitlmine, n_data_genes = 300, universe = "FANTOM_L")
 }
}
}
