% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_enrichment.R
\name{tidy_gp_results}
\alias{tidy_gp_results}
\title{Tidy functional enrichment results}
\usage{
tidy_gp_results(gp_result)
}
\arguments{
\item{gp_result}{-- The return object of gp_enrichment()}
}
\value{
A tibble
}
\description{
Returns a tidy tibble of the gprofiler output
}
\examples{
\dontrun{
if(interactive()){
 data_genes |>
   gp_enrichment(0.001, "WP") |>
   tidy_gp_results()
 }
}
}
