% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ards_literature_overlap.R
\name{unidentified_table}
\alias{unidentified_table}
\title{Create a table of genes identified in an alternative list not identified by ARDS MAIC}
\usage{
unidentified_table(data_genes, data_alternative, biolitmine = TRUE)
}
\arguments{
\item{data_genes}{-- Input in the format of \code{data_genes}}

\item{data_alternative}{-- Dataframe where the only column is gene symbols and titled gene,
unless is \code{data_biolitmine}}

\item{biolitmine}{-- TRUE or FALSE -- Default = TRUE -- if comparing BioLitMine result using
\code{data_biolitmine}}
}
\value{
An html table (biolitmine) or a tibble (other source)
}
\description{
Create an interactive html table of genes found in an alternative list but not by
ARDS MAIC.
}
\details{
Primary use is to compare ARDS MAIC with the results of a BioLitMine search for the MeSH
term "Respiratory Distress Syndrome, Adult". A secondary use case is to compare ARDS MAIC
with MAIC analyses performed in different conditions e.g., our pervious MAIC of SARS-CoV-2
whole-genome studies.
}
\examples{
\dontrun{
if(interactive()){
 unidentified_table(data_genes, data_biolitmine, biolitmine = TRUE)
 }
}
}
