% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_contributions_chord_diagrams.R
\name{contributions_chord_bymethod}
\alias{contributions_chord_bymethod}
\title{Contributions chord by method}
\usage{
contributions_chord_bymethod(data_contributions, data_study)
}
\arguments{
\item{data_contributions}{Data frame in the format of \code{ARDSMAICR::data_contributions}}

\item{data_study}{Data frame in the format of \code{ARDSMAICR::data_study}}
}
\value{
An html chord diagram
}
\description{
Creates an interactive html chord diagram of contributions to MAIC grouped by
method.
}
\details{
Input columns for \code{data_contributions} should be:
\itemize{
\item \code{study} - Study ID - chr
\item \code{raw_information} - Raw information. Sum of gene scores for each study - dbl
\item \code{information} - Information relative to the sum of all gene scores - dbl
\item \code{raw_contribution} - Raw contribution. Sum of gene scores for each study only where a score
contributes to the overall MAIC score - dbl
\item \code{contribution} - Contribution relative to the sum of all such scores - dbl
}

Input columns for \code{data_study} should be:
\itemize{
\item \code{id} - Integer 1 to n studies - dbl
\item \code{First_author} - First author family name - chr
\item \code{Article_title} - Article title - chr
\item \code{Year} - Year of publication - dbl
\item \code{Journal} - Journal - chr
\item \code{DOI} - Digital object identifier - dbl
\item \code{PMID} - PubMed ID - dbl
\item \code{uID} - Unique ID. Format is \verb{First_Author Year PMID} - chr
\item \code{Method} - Study method e.g., "GWAS" - chr
\item \code{Technology} - Technology used e.g., "Microarray" - chr
\item \code{Tissue} - Tissue type sampled e.g., "BALF" - chr
\item \code{Cell} - Cell type sampled e.g., "Neutrophils" - chr
\item \code{Focus} Study focus e.g., "Susceptibility" - chr
\item \code{ARDS_pts} - Total number of patients with ARDS included in study - dbl
\item \code{ARDS_definition} - Definition of ARDS used in study - chr
\item \code{List_available} - Was the gene list associated with the study retrievable - lgl
}
}
\examples{
\dontrun{
if(interactive()){
 contributions_chord_bymethod(ARDSMAICR::data_contributions, ARDSMAICR::data_study)
 }
}
}
