% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_contributions_chord_diagrams.R
\name{contributions_chord_bystudy}
\alias{contributions_chord_bystudy}
\title{Create a chord diagram of contributions by study}
\usage{
contributions_chord_bystudy(data)
}
\arguments{
\item{data}{-- Input in the format of \code{data_contributions}}
}
\value{
An html chord plot
}
\description{
Creates an interactive html plot of contributions by study
}
\examples{
\dontrun{
if(interactive()){
 contributions_chord_bystudy(data_genes)
 }
}
}
