% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_methods_and_gene_support.R
\name{count_genes}
\alias{count_genes}
\title{Count genes}
\usage{
count_genes(data)
}
\arguments{
\item{data}{-- Input in the format of \code{data_genes}}
}
\value{
An integer
}
\description{
Return the total number of unique genes or SNPs across all gene lists
}
\examples{
\dontrun{
if(interactive()){
 count_genes(data_genes)
 }
}
}
