% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_table.R
\name{gene_table}
\alias{gene_table}
\title{Table of genes}
\usage{
gene_table(data)
}
\arguments{
\item{data}{Data frame in the format of \code{ARDSMAICR::data_genes}}
}
\value{
An html table
}
\description{
Creates a searchable html table of genes ranked by MAIC score.
}
\details{
Input columns for \code{data_genes} should be (this is the standard output of the MAIC algorithm):
\itemize{
\item \code{gene} - HGNC gene symbol - chr
\item 1...
\item \code{uID} - Study unique identifier. Column contains study specific gene score - dbl
\item n...
\item \code{maic_score} - MAIC score for gene - dbl
\item \code{contributors} - Studies contributing to MAIC score by method - chr
}
}
\examples{
\dontrun{
if(interactive()){
 gene_table(ARDSMAICR::data_genes)
 }
}
}
