% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_methods_and_gene_support.R
\name{genes_per_list}
\alias{genes_per_list}
\title{Count genes per gene list}
\usage{
genes_per_list(data)
}
\arguments{
\item{data}{-- Input in the format of \code{data_genes}}
}
\value{
A tibble
}
\description{
Return the number of unique genes or SNPs identified in each gene list
}
\examples{
\dontrun{
if(interactive()){
 print(genes_per_list(data_genes), n = 50)
 }
}
}
