% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_enrichment.R
\name{gp_plot}
\alias{gp_plot}
\title{Functional enrichment plot}
\usage{
gp_plot(gp_result, is_interactive = TRUE)
}
\arguments{
\item{gp_result}{The return object of \code{gp_enrichment()}}

\item{is_interactive}{Interactive plot -- Default = TRUE}
}
\value{
A Manhattan plot
}
\description{
Creates a Manhattan plot of functional enrichment results.
}
\examples{
\dontrun{
if(interactive()){
 gp_enrichment(ARDSMAICR::data_genes, p_threshold = 0.001, source = c("REAC", "WP")) |>
 gp_plot(is_interactive = FALSE)
 }
}
}
