% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_prioritisation.R
\name{inflection_point_plot}
\alias{inflection_point_plot}
\title{Create a plot of the inflection point}
\usage{
inflection_point_plot(data)
}
\arguments{
\item{data}{-- Input in the format of \code{data_genes}}
}
\value{
A point density plot
}
\description{
Create a plot of genes ranked by maic score with the inflection point annotated and
point density of genes illustrated
}
\examples{
\dontrun{
if(interactive()){
 inflection_point_plot(data_genes)
 }
}
}
