% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_prioritisation.R
\name{inflection_point_plot}
\alias{inflection_point_plot}
\title{Plot the inflection point}
\usage{
inflection_point_plot(data, first_break = 1000, increment = 500)
}
\arguments{
\item{data}{Data frame in the format of \code{ARDSMAICR::data_genes}}

\item{first_break}{Integer value for first break on x-axis after inflection point -- Default = 1000}

\item{increment}{Integer value for size of increment for subsequent breaks on x-axis -- Default = 500}
}
\value{
A point density plot
}
\description{
Creates a plot of genes ranked by MAIC score with the inflection point annotated and
the point density of genes illustrated.
}
\details{
Input columns for \code{data_genes} should be (this is the standard output of the MAIC algorithm):
\itemize{
\item \code{gene} - HGNC gene symbol - chr
\item 1...
\item \code{uID} - Study unique identifier. Column contains study specific gene score - dbl
\item n...
\item \code{maic_score} - MAIC score for gene - dbl
\item \code{contributors} - Studies contributing to MAIC score by method - chr
}
}
\examples{
\dontrun{
if(interactive()){
 inflection_point_plot(ARDSMAICR::data_genes, first_break = 2000, incremnt = 1000)
 }
}
}
