% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_methods_and_gene_support.R
\name{lists_per_gene}
\alias{lists_per_gene}
\title{Count gene lists by gene}
\usage{
lists_per_gene(data)
}
\arguments{
\item{data}{Data frame in the format of \code{ARDSMAICR::data_genes}}
}
\value{
A tibble
}
\description{
For each unique gene or SNP returns the number of lists that gene or SNP is
found in.
}
\details{
Input columns for \code{data_genes} should be (this is the standard output of the MAIC algorithm):
\itemize{
\item \code{gene} - HGNC gene symbol - chr
\item 1...
\item \code{uID} - Study unique identifier. Column contains study specific gene score - dbl
\item n...
\item \code{maic_score} - MAIC score for gene - dbl
\item \code{contributors} - Studies contributing to MAIC score by method - chr
}
}
\examples{
\dontrun{
if(interactive()){
 ARDSMAICR::data_genes |>
   lists_per_gene() |>
   dplyr::filter(.data$gene == "IL6")
 }
}
}
