% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_methods_upset_plot.R
\name{methods_upset}
\alias{methods_upset}
\title{Methods upset plot}
\usage{
methods_upset(data)
}
\arguments{
\item{data}{Data frame in the format of \code{data_genes}}
}
\value{
An upset plot
}
\description{
Creates an upset plot of genes, methods, MAIC scores, and their intersections.
}
\examples{
\dontrun{
if(interactive()){
 methods_upset(data_genes)
 }
}
}
