% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ards_literature_overlap.R
\name{overlap_venn}
\alias{overlap_venn}
\title{Create a venn diagram of ARDS MAIC and alternative gene list overlap}
\usage{
overlap_venn(data_genes, data_alternative, biolitmine = TRUE)
}
\arguments{
\item{data_genes}{-- Input in the format of \code{data_genes}}

\item{data_alternative}{-- Dataframe where the only column is gene symbols and titled gene,
unless is \code{data_biolitmine}}

\item{biolitmine}{-- TRUE or FALSE -- Default = TRUE -- if comparing BioLitMine result using
\code{data_biolitmine}}
}
\value{
Venn diagram
}
\description{
Create a venn diagram of the overlap between ARDS MAIC genes and genes found in an
alternative list.
}
\examples{
\dontrun{
if(interactive()){
 overlap_venn(data_genes, data_biolitmine, biolitmine = TRUE)
 }
}
}
